refactor: Product 도메인 분리
