+++ src/main/java/cart/domain/Product.java
@@ -1,13 +1,13 @@
-package cart.controller.dto;
+package cart.domain;
 
-public class ProductDto {
+public class Product {
 
     private final Long id;
     private final String name;
     private final String imageUrl;
     private final int price;
 
-    public ProductDto(final long id, final String name, final String imageUrl, final int price) {
+    public Product(final long id, final String name, final String imageUrl, final int price) {
         this.id = id;
         this.name = name;
         this.imageUrl = imageUrl;
