+++ src/main/resources/templates/index.html
@@ -27,7 +27,7 @@
             <div class="product-info">
                 <div class="product-desc">
                     <p class="product-name" th:text="${product.name}"></p> <!-- 상품 이름 -->
-                    <p class="product-price" th:text="${product.price}"></p> <!-- 상품 가격 -->
+                    <p class="product-price" th:text="${product.price} + '원'"></p> <!-- 상품 가격 -->
                 </div>
                 <button type="submit" class="product-btn" th:onclick="addCartItem([[${product.id}]])">담기</button>
             </div>
